исправил расположение файлов
