+++ src/resources/strings/countries_binary_mode_output.txt
