+++ src/resources/strings/countries_text_mode_output.txt
