+++ src/ru/avalon/java/j20/labs/tasks/Task2.java
@@ -1,16 +1,16 @@
-package ru.avalon.java.j20.labs.tasks;
+package src.ru.avalon.java.j20.labs.tasks;
 
-import ru.avalon.java.j20.labs.Task;
+import src.ru.avalon.java.j20.labs.Task;
 
 import java.io.*;
 
 /**
- * Задание №1
+ * Задание №2
  *
  * <p>Тема: "Потоковый ввод-вывод. Чтение и запись данных
- * в двоичном режиме".
+ * в текстовом режиме".
  */
-public class Task1 implements Task {
+public class Task2 implements Task {
 
     /**
      * {@inheritDoc}
@@ -18,33 +18,33 @@ public class Task1 implements Task {
     @Override
     public void run() throws IOException {
         File input = new File("assets/countries.txt");
-        File output = new File("countries_binary_mode_output.txt");
+        File output = new File("src/resources/strings/countries_text_mode_output.txt");
         String text = read(input);
         write(output, text);
 //        System.out.println(text);
 
         /*
-         * TODO(Студент): Выполнить задание №1
+         * TODO(Студент): Выполнить задание №2
          *
          * 1. Реализовать метод read.
          *
          *    При чтении файла следует пользоваться типами данных:
-         *    InputStream и FileInputStream.
+         *    Reader, FileReader.
          *
-         *    Для сохранениня прочитанных данных следует пользоваться
-         *    классом ByteArrayOutputStream.
+         *    Для сохранения прочитанных данных следует пользоваться
+         *    классом StringBuilder.
          *
          * 2. Реализовать метод write.
          *
          *    При реализации метода следует пользоваться типами данных:
-         *    OutputStream и FileOutputStream.
+         *    Writer и FileWriter.
          *
          * 3. С использованием отладчика проверить корректность работы программы.
          */
     }
 
     /**
-     * Выполняет чтение указанного файла в двоичном режиме.
+     * Выполняет чтение указанного файла в текстовом режиме.
      *
      * <p>Весь текст файла возвращается в виде одного
      * экземпляра типа {@link String}.
@@ -54,19 +54,19 @@ public class Task1 implements Task {
      * @throws IOException в случае ошибок ввода-вывода.
      */
     private String read(File file) throws IOException {
-        try (InputStream input = new FileInputStream(file);
-             ByteArrayOutputStream bos = new ByteArrayOutputStream()) {
-            byte[] buffer = new byte[5];
-            int data;
-            while ((data = input.read(buffer)) > 0) {
-                bos.write(buffer, 0, data);
+        try (Reader reader = new FileReader(file)) {
+            StringBuilder builder = new StringBuilder();
+            char[] buffer = new char[10];
+            int len;
+            while((len = reader.read(buffer)) != -1) {
+                builder.append(buffer, 0, len);
             }
-            return bos.toString();
+            return builder.toString();
         }
     }
 
     /**
-     * Выполняет запись текстоых данных в файл в двоичном
+     * Выполняет запись текстоых данных в файл в текстовом
      * режиме.
      *
      * @param file файл
@@ -74,8 +74,8 @@ public class Task1 implements Task {
      * @throws IOException в случае ошибок ввода-вывода.
      */
     private void write(File file, String text) throws IOException {
-        try (OutputStream output = new FileOutputStream(file)) {
-            output.write(text.getBytes());
+        try(Writer writer = new FileWriter(file)) {
+            writer.write(text);
         }
     }
 }
