+++ src/ru/avalon/java/j20/labs/tasks/Task3.java
@@ -1,6 +1,6 @@
-package ru.avalon.java.j20.labs.tasks;
+package src.ru.avalon.java.j20.labs.tasks;
 
-import ru.avalon.java.j20.labs.Task;
+import src.ru.avalon.java.j20.labs.Task;
 
 import java.io.*;
 import java.util.*;
@@ -18,7 +18,7 @@ public class Task3 implements Task {
     @Override
     public void run() throws IOException {
         File input = new File("assets/countries.txt");
-        File output = new File("countries_buffered_mode_output.txt");
+        File output = new File("src/resources/strings/countries_buffered_mode_output.txt");
         Collection<String> lines = read(input);
         write(output, lines);
